Refactor API structure and move business logic to services
Reorganized the backend by moving API endpoints to the 'api/endpoints' directory and extracting business logic into dedicated 'services' modules. Removed the old 'routers' and utility files, updated configuration management to 'core/config.py', and improved code modularity and maintainability. Email and template operations are now handled in 'email_service.py' and 'template_service.py', while database utilities reside in 'utils/db_utils.py'.
