+++ backend/app/api/__init__.py
@@ -1,0 +1,3 @@
+"""
+API package for all API related endpoints.
+"""
