+++ backend/app/api/endpoints/__init__.py
@@ -1,0 +1,3 @@
+"""
+API endpoints module.
+"""
