+++ backend/app/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core module for essential application functionality.
+"""
