--- backend/app/routers/__init__.py
@@ -1,4 +1,0 @@
-"""
-Module init file for routers package.
-Makes the directory a proper Python package.
-"""
