+++ backend/app/services/__init__.py
@@ -1,0 +1,3 @@
+"""
+Services module for business logic implementations.
+"""
